requirements.txt
flask==3.0.2
gTTS==2.5.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0 